[posix] relocate PosixMountProvider from platform/linux to platform/posix
